+++ src/app/(dashboard)/(pages)/connections/page.tsx
@@ -1,4 +1,4 @@
-import { CONNECTIONS } from "@/lib/constants";
+import { CONNECTIONS } from "@/lib/misc/constants";
 import ConnectionCard from "./_components/connection-card";
 
 type ConnectionsProps = {
