+++ src/app/(dashboard)/(pages)/settings/_components/profile-picture.tsx
@@ -6,7 +6,7 @@ import { Button } from "@/components/ui/button";
 import { X } from "lucide-react";
 import { User } from "@prisma/client";
 import UploadCareButton from "@/components/shared/uploadcare-button";
-import { env } from "@/lib/env";
+import { env } from "@/lib/config/env";
 
 type ProfilePictureProps = {
   userImage: string | null;
