+++ src/app/_components/infinite-moving-cards.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/lib/client/utils";
 import Image from "next/image";
 import { useCallback, useEffect, useRef, useState } from "react";
