+++ src/app/api/clerk-webhook/route.ts
@@ -1,4 +1,4 @@
-import { db } from "@/lib/db";
+import { db } from "@/lib/database";
 import { NextResponse } from "next/server";
 
 export async function POST(req: Request) {
