+++ src/components/shared/card-3d.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/lib/client/utils";
 import React, {
   createContext,
   useState,
