+++ src/components/shared/uploadcare-button.tsx
@@ -3,7 +3,7 @@
 import { useEffect, useRef, useState } from "react";
 import * as UC from "@uploadcare/file-uploader";
 import Spinner from "@/components/shared/spinner";
-import { createEventListener } from "@/lib/utils";
+import { createEventListener } from "@/lib/client/utils";
 
 type UploadCareButtonProps = {
   onUpload: (cdnUrl: string) => void;
