+++ src/components/ui/dropdown-menu.tsx
@@ -4,7 +4,7 @@ import * as React from "react";
 import * as DropdownMenuPrimitive from "@radix-ui/react-dropdown-menu";
 import { Check, ChevronRight, Circle } from "lucide-react";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/lib/client/utils";
 
 const DropdownMenu = DropdownMenuPrimitive.Root;
