+++ src/components/ui/form.tsx
@@ -12,7 +12,7 @@ import {
   useFormContext,
 } from "react-hook-form";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/lib/client/utils";
 import { Label } from "@/components/ui/label";
 
 const Form = FormProvider;
